requirements.txt
fastapi
python-multipart
uvicorn
boto3
python-dotenv